Refactor tests
Current integration tests categorization doesn't scale well for the
amount of tests cases we have. I made the following changes to address
it:

1. Split `init_replication` test into `removals`, `insertion`, `spawn`
   and `despawn`.
2. Split insertion and removal test with multiple components into
   separate tests.
3. Thanks to code coverage I found a missing test case for spawning an
   entity from a mapped component, added.
